+++ tests/conftest.py
@@ -1,42 +1,12 @@
-import importlib
 from pathlib import Path
 from typing import Type
 
 from lxml import etree
 
-from xsdata.formats.dataclass import utils
 from xsdata.formats.dataclass.parsers import JsonParser
 from xsdata.formats.dataclass.parsers import XmlParser
 from xsdata.formats.dataclass.serializers import JsonSerializer
 from xsdata.formats.dataclass.serializers import XmlSerializer
-from xsdata.utils import text
-
-
-def load_class(output, clazz_name):
-    search = "Generating package: "
-    start = len(search)
-    packages = [line[start:] for line in output.split("\n") if line.startswith(search)]
-
-    for package in reversed(packages):
-        try:
-            module = importlib.import_module(package)
-            return getattr(module, clazz_name)
-        except (ModuleNotFoundError, AttributeError):
-            pass
-
-    raise ModuleNotFoundError(f"Class `{clazz_name}` not found.")
-
-
-def read_root_name(path: Path) -> str:
-    try:
-        recovering_parser = etree.XMLParser(recover=True)
-        tree = etree.parse(str(path), parser=recovering_parser)
-        local_name = etree.QName(tree.getroot().tag).localname
-        return text.pascal_case(utils.safe_snake(local_name, "Type"))
-    except etree.XMLSyntaxError:
-        return ""
-    except OSError:
-        return ""
 
 
 def validate_bindings(schema: Path, clazz: Type):
