+++ tests/integration/test_books.py
@@ -4,8 +4,8 @@ from click.testing import CliRunner
 
 from tests import fixtures_dir
 from tests import root
-from tests.conftest import load_class
 from xsdata.cli import cli
+from xsdata.utils.testing import load_class
 
 os.chdir(root)
