+++ tests/integration/test_calculator.py
@@ -6,7 +6,6 @@ from click.testing import CliRunner
 
 from tests import fixtures_dir
 from tests import root
-from tests.conftest import load_class
 from tests.fixtures.calculator import CalculatorSoapAdd
 from tests.fixtures.calculator import CalculatorSoapAddOutput
 from xsdata.cli import cli
@@ -14,6 +13,7 @@ from xsdata.formats.dataclass.client import Client
 from xsdata.formats.dataclass.client import Config
 from xsdata.formats.dataclass.serializers import XmlSerializer
 from xsdata.formats.dataclass.transports import DefaultTransport
+from xsdata.utils.testing import load_class
 
 os.chdir(root)
